Entities have friction, improved Input
Input "Held" now returns true as soon as a key or button is pressed.

Was going to try time invariant updates for entities, but Haxe's support
for timers are not so good. There just wasn't an accurate timer that
would work on all platforms. Sys.time(0) doesn't work on html5.