Entity lists and spawning
Also moved camera and level to their own classes.

Adding new entities presents a problem. I would like to have a
recycleable list, where deactivated entities just lie dormant, and new
entities can copy themselves into the deactivated entity and reactivate.

The downside is, this would mean that the entity type would have to be a
jack of all trades, being able to transform into any entity just by
altering it's internal data. It would create a huge, unweildy entity
class.

So currently, I'm just creating new objects, appending them to the
entity list, and destroying the inactive entities. I'm worried this will
cause garbage accumulation. But maybe modern computers can just deal
with the overhead.

Another thing to fix is the way the shot handles the level borders. It
needs to gain knowledge of the level borders, so it can deal with it
it's own way (deactivate itself), as opposed to the generic way of just
stopping at the borders.