Stress testing and adding an option to change framerate regulation method
The results of stress testing are inconclusive. The non-recycling method
seems to use slightly more memory than the recycling method, but both
end up with similar framerates at high stress levels (4000 entities).
Will probably end up removing the recycling method for default entity
creation, but might use a recycling list for specific kinds of entities
(like debris or explosions or particle effects).

Two different ways of regulating frame rate: Enterframe method and timer
method. The timer method will always keep it at 62 fps or 58 fps. Sadly,
it cannot do exactly 60. The delay between frames is determined by
number of milliseconds, and it must be a whole number, can't be
fractions. So 16 = 62 fps. 17 = 58 fps. The enterframe method uses the
exact refresh rate of the monitor, and results in smoother motion. But
it depends on the monitor refresh rate. So a 100 hz refresh rate will
cause the framerate to be 100 fps. That will affect physics, unless the
physics calculations takes this into account (currently it does, but im
unsure of accuracy). It will also complicate demo recording and playback
(a future feature I wish to include).