Created Hitshape, moved collision logic to hitshape
Moved the minutia of collision logic to HitShape. Hitshape contains and array of points, definining the shape of the entity's solid body. This will allow for diagonal walls in level tiles.

I'm liking the way this is being handled. A little unsure about the static "movefraction" and "pushoutX,Y" stuff, but I'm not sure how else to handle it.

I will add surface bouncing next, so it doesn't just stick to surfaces like flypaper.