Fixed door bug
The problem was that it was calling SwitchRoomPlayerPosX on the NEW room, but the new room didn't have the necessary information. Now calling it will get it from the old room.

Switching door orientation won't work, even if it looks fine now. It needs more specific rotation information to work, and that's too much work.