Fixed bug where player can get stuck between rooms
Bug existed where player could be stuck perpetually switching rooms. It would make the player spawn in the door entrance trigger, and attempting to move into the room caused the player to switch back to previous room, also behind the room switching trigger.

Fixed by making the player enter the room 1 unit away from the door, according to the direction of the door.