Level tile rendering optimization test
Using large numbers of black tiles seems to cause lag. Using large Shapes to mask the background seems to cause lag too. I had an idea to use bitmaps. Right now it's a pre-created bitmap, but in the future it should be dynamically generated based on the "0"s in the room's tilemap.