Added HUD
Heads up display shows 3 bars, representing the status of the player. Health, Shield, and Boost.

The exact health, shield and boost recharging mechanism is not done yet. This is a simplified model, for demonstrating the HUD bars. Shield recharging is going to be much more complicated. Health may not recharge at all. Boost may remain the same.